Be more permissive in gem deps
The ~> operator accepts new versions based on the last provided dot
version. So ~> 0.3.0 would allow 0.3.2 but not 0.4.0. This is often
problematic for keeping multiple deps updated as minor updates get
blocked.

A common Ruby idiom is to specify the MAJOR.MINOR in semver's
MAJOR.MINOR.PATCH. It is generally viewed as relatively "safe" to accept
new minor releases as they _should_ be backwards compatible per semver.